feat: working draft for detailed single sample handler

+++ transcriptomics_data_service/models.py
@@ -76,7 +76,7 @@ class GeneExpression(BaseModel):
     gene_code: str = Field(..., min_length=1, max_length=255, description="Feature identifier")
     sample_id: str = Field(..., min_length=1, max_length=255, description="Sample identifier")
     experiment_result_id: str = Field(..., min_length=1, max_length=255, description="ExperimentResult identifier")
-    raw_count: int = Field(..., ge=0, description="The raw count for the given feature")
+    raw_count: float = Field(..., ge=0, description="The raw count for the given feature")
     tpm_count: float | None = Field(None, ge=0, description="TPM normalized count")
     tmm_count: float | None = Field(None, ge=0, description="TMM normalized count")
     getmm_count: float | None = Field(None, ge=0, description="GETMM normalized count")
